brain/requirements.txt
# ═══════════════════════════════════════════════════════════════
# Prime PenTrix Brain API - Lightweight RAG Dependencies
# Where Penetration Testing Meets Intelligence
#
# LIGHTWEIGHT approach: BM25 + OpenAI API embeddings
# Total: ~100MB (NOT 500MB with sentence-transformers)
# ═══════════════════════════════════════════════════════════════

# ── Core Framework ──────────────────────────────────────────
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.12       # File upload support
pydantic==2.9.2
pydantic-settings==2.5.2

# ── Database (PostgreSQL + pgvector) ────────────────────────
psycopg2-binary==2.9.9         # Sync PostgreSQL driver (lighter than asyncpg)

# ── AI Embeddings (via API, NOT local models) ───────────────
httpx==0.27.0                  # Async HTTP client for OpenAI API
python-dotenv==1.0.1           # Environment variable loading

# ── Document Processing ────────────────────────────────────
pypdf==4.3.1                   # PDF text extraction (lightweight)

# ── Text Search ────────────────────────────────────────────
rank-bm25==0.2.2               # BM25 keyword search (50KB!)
nltk==3.9.1                    # Tokenization for BM25

# ── Utilities ──────────────────────────────────────────────
numpy==1.26.4                  # Array operations for embeddings

# ═══════════════════════════════════════════════════════════════
# REMOVED (too heavy, not needed):
# ❌ sentence-transformers  (300MB + pulls torch)
# ❌ torch                  (2GB)
# ❌ transformers           (200MB)
# ❌ asyncpg                (replaced with psycopg2-binary)
# ❌ pgvector[asyncpg]      (using raw SQL for pgvector)
# ❌ chromadb               (using pgvector directly)
# ═══════════════════════════════════════════════════════════════
